+++ apps/web/src/modules/registry.tsx
@@ -88,6 +88,7 @@ export const MODULE_COMPONENTS: Record<ModuleCode, LazyExoticComponent<Component
   XQL: mod(() => import('./ExecQualityModule'), 'ExecQualityModule'),
   START: mod(() => import('./StartModule'), 'StartModule'),
   SYS: mod(() => import('./SysModule'), 'SysModule'),
+  KEYS: mod(() => import('./KeysModule'), 'KeysModule'),
   RHEAT: mod(() => import('./PortfolioRiskModule'), 'PortfolioRiskModule'),
   EXP: mod(() => import('./ExposureModule'), 'ExposureModule'),
   PBETA: mod(() => import('./PortfolioBetaModule'), 'PortfolioBetaModule'),